clear all node_modules if modified

+++ src/lib/git.js
@@ -26,14 +26,10 @@ export function captureUntracked(repoRoot) {
 }
 
 export function cleanWorkingDirectory(repoRoot, links, modified) {
+  if (!repoRoot) throw new Error('No root directory specified')
   execSync('git reset --hard HEAD', { cwd: repoRoot, encoding: 'utf8' })
   execSync('git clean -fd', { cwd: repoRoot, encoding: 'utf8' })
-
-  for (const { package: pkgPath } of modified) {
-    try {
-      fs.rmSync(pkgPath, { recursive: true, force: true })
-    } catch {}
-  }
+  if (modified.length) execSync('rm -rf ./node_modules', { cwd: repoRoot, encoding: 'utf8' })
 }
 
 export function applyPatch(patch, repoRoot) {
